react form stable

+++ src/features/petForm/lib/prisma_actions.ts
@@ -5,45 +5,69 @@ import { revalidatePath } from "next/cache";
 import prisma from "@/shared/lib/prismaDB";
 import { sleep } from "@/shared/lib/hooks/utils";
 import { Pet } from "@/generated/prisma";
-import { PetFormData } from "@/entities/dashboard/model/types";
+import { TPetForm, petFormSchema, petIdSchema } from "@/shared/lib/validation";
+import { safeParse } from "zod/v4/core";
 
 export const getPets = async () => {
   return await prisma.pet.findMany();
 };
-export const addPet = async (pet: PetFormData) => {
-  sleep(1000);
+export const addPet = async (pet: unknown) => {
+  await sleep(1000);
+
+  // before there was an error, where age was not a number, because typing wasn't working on server
+  const validatedPet = petFormSchema.safeParse(pet);
+  if (!validatedPet.success) {
+    return { error: "Invalid pet data", id: null };
+  }
 
   try {
-    return await prisma.pet.create({
-      data: pet,
+    const result = await prisma.pet.create({
+      data: validatedPet.data,
     });
+
+    revalidatePath("/app", "layout");
+    return { error: null, id: result.id };
   } catch (error) {
-    throw new Error("Could not add pet");
+    return { error: "Could not add pet", id: null };
   }
 };
 
-export const updatePet = async (id: Pet["id"], pet: PetFormData) => {
+export const updatePet = async (id: Pet["id"], pet: TPetForm) => {
   sleep(1000);
+
+  const validatedPet = petFormSchema.safeParse(pet);
+  const validatedId = petIdSchema.safeParse(id);
+  if (!validatedPet.success || !validatedId.success) {
+    return { message: "Invalid pet data" };
+  }
+
   try {
     await prisma.pet.update({
-      where: { id },
-      data: pet,
+      where: { id: validatedId.data },
+      data: validatedPet.data,
     });
   } catch (error) {
-    throw new Error("Could not update pet");
+    return { message: "Could not update pet" };
   }
   revalidatePath("/app", "layout");
 };
+
 export const deletePet = async (id: Pet["id"]) => {
   sleep(1000);
+
+  const validatedId = petIdSchema.safeParse(id);
+  if (!validatedId.success) {
+    return { message: "Invalid pet data" };
+  }
+
   try {
     await prisma.pet.deleteMany({
       where: {
-        id: id,
+        id: validatedId.data,
       },
     });
   } catch (error) {
-    throw new Error("Could not delete pet");
+    return { message: "Could not delete pet" };
   }
   revalidatePath("/app", "layout");
 };
